Refactor UserCard models
