+++ src/client/lib/authAPI/index.ts
@@ -1,10 +1,7 @@
 import { getAPIWithAuth } from '../fetchAPI'
 
-import {
-    // SHARED_APIFields__GetLifetimeStats,
-    SHARED_APIFields__GetUserCard,
-    SHARED_APIFields__LifetimeStats,
-} from '@/shared/models'
+import { SHARED_APIFields__LifetimeStats } from '@/shared/models'
+import { SHARED_APIFields__GetUserCard } from '@/shared/models/apiFields'
 
 /*
 Authenticated APIs (must provide access token)
