+++ src/shared/models/apiFields/index.ts
@@ -1,3 +1,5 @@
+import { SHARED_Model__UserCardData } from '../models/UserCard'
+
 export interface SHARED_APIFields__BASE {
     success: boolean
     error?: string
@@ -8,3 +10,7 @@ export interface SHARED_APIFields__GetToken extends SHARED_APIFields__BASE {
 }
 
 export type SHARED_APIFields__ValidateToken = SHARED_APIFields__BASE
+
+export interface SHARED_APIFields__GetUserCard extends SHARED_APIFields__BASE {
+    userCard?: SHARED_Model__UserCardData
+}
