+++ src/shared/models/models/UserCard.ts
@@ -1,0 +1,9 @@
+export interface SHARED_Model__UserCardData {
+    userId: string
+    name: string | undefined
+    email: string | undefined
+    followers: number
+    following: number
+    createdAt: string
+    totalRepos: number
+}
